Remove Tree, use clade count
Originally this code used the output of the V1 pipeline, which did not include deduped cladecounts.  Instead, to get the deduped count of an HV clade you'd count all the reads assigned to members of the clade.  But now we're reading V2 output we can just use the clade count directly. This means we can stop using human_virus_tree-2022-12.json, which isn't even necessarily using the same taxids as the V2 output. It also means we can delete a lot of tree-wrangling code.